Fix 317951 - Horizontal Frame on system by itself at end of page causes previous system to disappear
Will return the correct bottom y coordinate in case a SysStaff contains only a horizontal frame.
In this case the SysStaff doens't have a valid skyline and the initial (far too large) maximum value is returned.
This is solved by returning the height of the SysStaff in case the skyline is invalid (has no segments).